requirements.txt
pandas==1.2.4
PyYAML==5.4.1
typer==0.3.2
albumentations==1.0
dvc==2.0.17
hydra_core==1.1.0
loguru==0.5.3
mlflow==1.18.0
omegaconf==2.1.0
scikit_learn==0.24.2

